requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
ffmpeg-python==0.2.0
Flask==3.0.1
future==0.18.3
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.0.3
tqdm==4.66.2
urllib3==2.2.0
Werkzeug==3.0.1
